+++ src/main/kotlin/dev/thynanami/idea/typst/config/TypstCodeStyleSettings.kt
@@ -1,0 +1,10 @@
+package dev.thynanami.idea.typst.config
+
+import com.intellij.psi.codeStyle.CodeStyleSettings
+import com.intellij.psi.codeStyle.CustomCodeStyleSettings
+
+class TypstCodeStyleSettings(settings: CodeStyleSettings) : CustomCodeStyleSettings("TypstCodeStyleSettings", settings) {
+    @JvmField
+    @Suppress("PropertyName")
+    var PROSE_WRAP = false
+}